google_cloud_function/requirements.txt
hyperspy
numpy
matplotlib
pillow
scikit-image
google-cloud-storage>=1.38.0

./nexuslims